:truck: api: move save-jury-complementary-certif-course-result usecase to src
